Fix: Response types 수정, 불필요한 types 제거
